Added an example file to use struct for sibhfs.

+++ matlab/low/3d/clfids2hfid.m
@@ -3,4 +3,4 @@ function hfid = clfids2hfid(cid, lfid) %#codegen
 % HFID = CLFIDS2HFID(CID, LFID)
 % See also HFID2CID, HFID2LFID
 
-hfid = cid*8+lfid-1;
+hfid = cid*8+int32(lfid)-1;
